Got all the holes filled!
As far as I can tell flying around in the editor, it looks like all the holes in the terrain have been filled in 😊

A big part of the problem was that I did fully understand bitmasks, and I defined the one I'm using to determine if a block has been processed or not wrong. With that fixed, there was just a couple bugs with the rendering of Left/Right walls at the edges of a chunk.