More refactoring, fixed a bug
Making z a uint8 really threw things out of whack. When it decremented all the way to 0 and *wrapped around to 255* the index attempting to be retrieved was *way* outside the bounds of the array.

It's an int8 now; which works as expected. The slightly unfortunate side effect is that Chunks can now be a maximum of 128 blocks tall instead of 256.